Create CHANGELOG.md
When several people are working on a project, the probability that all of them are aware of all the changes is almost zero. Project management tools (usually) focus more on the tasks at hand than to record & register a development history. The changelog represents the current state of your code base and with each release, it allows all team members to be aware of exactly what new file or "thing" was built, changed, fixed or merged.